+++ example/deepseek_example.dart
@@ -8,7 +8,7 @@ void main() async {
     // 🔹 Create a chat completion -> https://api-docs.deepseek.com/api/create-chat-completion
     Completion response = await deepSeek.createChat(
       messages: [Message(role: "user", content: "Hello, how are you?")],
-      model: Models.chat,
+      model: Models.chat.name,
       options: {
         "temperature": 1.0,
         "max_tokens": 4096,
@@ -17,13 +17,15 @@ void main() async {
     print("Chat Response: ${response.text}");
 
     // 🔹 List available models
-    List<Models> models = await deepSeek.listModels();
+    List<String> models = await deepSeek.listModels();
     print("Available Models: $models");
 
     // 🔹 Get user balance
     Balance balance = await deepSeek.getUserBalance();
     print("User Balance: ${balance.info}");
+  } on DeepSeekException catch (e) {
+    print("Api returned error: ${e.statusCode}:${e.message}");
   } catch (e) {
-    print("Error: $e");
+    print("something unexpected happened: $e");
   }
 }